Changed to lower max value

+++ AngleSharp/Css/ValueConverters/GradientConverter.cs
@@ -105,7 +105,7 @@
 
         public Int32 MaxArgs
         {
-            get { return Int32.MaxValue; }
+            get { return UInt16.MaxValue; }
         }
     }
 }
