+++ AngleSharp/Css/ValueConverters/OrderedOptionsConverter.cs
@@ -16,7 +16,7 @@
 
         public Int32 MaxArgs
         {
-            get { return Int32.MaxValue; }
+            get { return UInt16.MaxValue; }
         }
 
         public Int32 MinArgs
