+++ AngleSharp/Extensions/ValueConverterExtensions.cs
@@ -86,7 +86,7 @@
             return validate(converter, null);
         }
 
-        public static IValueConverter<T[]> Many<T>(this IValueConverter<T> converter, Int32 min = 1, Int32 max = Int32.MaxValue)
+        public static IValueConverter<T[]> Many<T>(this IValueConverter<T> converter, Int32 min = 1, Int32 max = UInt16.MaxValue)
         {
             return new OneOrMoreValueConverter<T>(converter, min, max);
         }
